Moving added function from `Logic/Posts.php` to `Logic/ConsoleOutput/Posts.php`
